refactor: Make user_id 64bit
Allows, giving out user_id's like candy without risking exhaustion.

+++ src/client_net.rs
@@ -4161,7 +4161,7 @@ mod tests {
     use super::*;
     use std::time::{SystemTime, UNIX_EPOCH};
 
-    fn user(id: u32) -> UserId {
+    fn user(id: u64) -> UserId {
         UserId(id)
     }
